pyproject.toml
[project]
name = "imgs2xxx"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fpdf>=1.7.2",
    "pillow>=11.2.1",
]
